requirements.txt
yfinance
pandas
requests
python-dotenv
vaderSentiment
feedparser
twilio